+++ yapaghetti-application/src/main/java/kr/hs/entrydsm/yapaghetti/domain/document/api/CreateLocalDocumentPort.java
@@ -1,7 +1,7 @@
 package kr.hs.entrydsm.yapaghetti.domain.document.api;
 
-import kr.hs.entrydsm.yapaghetti.domain.document.api.dto.request.DomainCreateLocalDocumentRequest;
+import kr.hs.entrydsm.yapaghetti.domain.document.api.dto.request.DomainCreateDocumentRequest;
 
 public interface CreateLocalDocumentPort {
-    void execute(DomainCreateLocalDocumentRequest request);
+    void execute(DomainCreateDocumentRequest request);
 }
