+++ yapaghetti-application/src/main/java/kr/hs/entrydsm/yapaghetti/domain/document/api/dto/request/DomainCreateDocumentRequest.java
