+++ yapaghetti-application/src/main/java/kr/hs/entrydsm/yapaghetti/domain/document/usecase/CreateLocalDocumentUseCase.java
@@ -2,7 +2,7 @@ package kr.hs.entrydsm.yapaghetti.domain.document.usecase;
 
 import kr.hs.entrydsm.yapaghetti.annotation.UseCase;
 import kr.hs.entrydsm.yapaghetti.domain.document.api.CreateLocalDocumentPort;
-import kr.hs.entrydsm.yapaghetti.domain.document.api.dto.request.DomainCreateLocalDocumentRequest;
+import kr.hs.entrydsm.yapaghetti.domain.document.api.dto.request.DomainCreateDocumentRequest;
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.Document;
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.DocumentType;
 import kr.hs.entrydsm.yapaghetti.domain.document.spi.CommandDocumentPort;
@@ -20,7 +20,7 @@ public class CreateLocalDocumentUseCase implements CreateLocalDocumentPort {
     private final CommandDocumentPort commandDocumentPort;
 
     @Override
-    public void execute(DomainCreateLocalDocumentRequest request) {
+    public void execute(DomainCreateDocumentRequest request) {
         User user = documentQueryUserPort.queryUserById(documentSecurityPort.getCurrentUserId());
 
         commandDocumentPort.saveDocument(
